requirements.txt
allure-pytest==2.13.2
pytest==7.4.0
requests==2.31.0
selenium==4.10.0
faker==18.11.1
pytest-html==4.0.0
flake8==6.1.0
autopep8==2.0.4